add member fns
