+++ src/transcode/export.ts
@@ -1,3 +1,4 @@
 export * from './decode';
 export * from './encode';
 export * from './utils';
+export * from './interface';
