+++ src/utils/event.ts
@@ -1,31 +1,7 @@
 import createKeccakHash from 'keccak';
-import { EventFragment, Result } from 'ethers/lib/utils';
-import { WarpInterface } from '../transcode/interface';
 import { toUintOrFelt } from './export';
 
 export type EventItem = { data: string[]; keys: string[]; order?: number };
-
-export function decodeWarpEvent(fragment: EventFragment | string, warpEvent: EventItem): Result {
-  warpEvent = join248bitChunks(warpEvent); // reverse 248 bit packing
-
-  // Remove leading 0x from each element and join them
-  const data = `0x${warpEvent.data.map((x) => x.slice(2)).join('')}`;
-
-  const result = new WarpInterface(fragment).decodeEventLog(fragment, data, warpEvent.keys);
-  return result;
-}
-
-export function encodeWarpEvent(fragment: EventFragment, values: argType[], order = 0): EventItem {
-  const { data, topics }: { data: string; topics: string[] } = new WarpInterface(
-    fragment,
-  ).encodeEventLog(fragment, values);
-
-  const topicFlatHex = '0x' + topics.map((x) => x.slice(2).padStart(64, '0')).join('');
-  const topicItems248: string[] = splitInto248BitChunks(topicFlatHex);
-  const dataItems248: string[] = splitInto248BitChunks(data);
-
-  return { order, keys: topicItems248, data: dataItems248 };
-}
 
 export function splitInto248BitChunks(data: string): string[] {
   if (data.startsWith('0x')) data = data.slice(2);
